requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Vector Database
chromadb==0.4.18
# pinecone-client==3.0.0  # Uncomment for production
# qdrant-client==1.7.0    # Alternative

# Graph Database
neo4j==5.14.1

# Structured Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4

# Cache
redis==5.0.1
hiredis==2.2.3

# LLM & Embeddings
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
sentence-transformers==2.2.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
unstructured==0.11.8
pdfplumber==0.10.3

# Text Processing
tiktoken==0.5.2
spacy==3.7.2
nltk==3.8.1

# Utilities
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tqdm==4.66.1
python-multipart==0.0.6

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.1
flake8==6.1.0
isort==5.13.2